Fix #454 - Improve Cosign logout process
 * Defer terminating the session until logout is confirmed
 * Purge Cosign cookie upon logout confirmation

There is a new route, `/logout_now`, which triggers the confirmation and
redirects to the front page. It is never actually seen by the user
because it serves only to terminate sessions and redirect, but it does
no harm if a user visits the URL directly.

This change is clunky to test for two reasons:

First, Devise triggers a login on any presence of the remote user
header. Because the Rack middleware that generates fake headers for easy
login in development is unconditional and requires restart to disable,
logging out results in an immediate login on the front page. Working
around this requires setting up a reverse proxy and disabling the
middleware.

Second, to test the full two-stage logout, an actual Cosign config is
required. This means that it is really only practical to test in staging
on a real hostname.