Fixes #454 - Delete cookie with proper flags
The problem with the other versions of the cookie deletion was that the
Cosign cookie is a "host-only" cookie. This is accomplished by not
setting the domain at all. The first attempt introduced the domain,
which made everything else fail. This is the correct approach, which
expires the host-only cookie (no domain sent) at the path of /. This
matches the existing cookie, deleting it as expected. The secure flag is
irrelevant because we are sending a new cookie, which expires
immediately.