Provide a specialized umich clamav setup to use the daemon
This wholesale replaces the use of gem `clamav` in production in lieu
of `clamav-client`, which connects to the running daemon and streams
the file to it 4k at a time.

The AV scan only reports an infection when it actively finds an
infection, which is how the [default virus checker](https://github.com/samvera/hydra-works/blob/68ffcda8025e76699f0f2c2bf1f42ee861579d34/lib/hydra/works/services/virus_checker_service.rb#L19) works.

Use of clamav is still relegated to only the `production` group.