add provenance logging version 1 features (#815)
A possibility for the future is to write to the database and not a log file.  This way we can access the audit trail for works and collections.  So if we have a table that has:
AuditTable
 - id ( could be work or collection if we know they will never share an id )
 - date of change
 - msg 

When an admin user goes to the show page of a work or a collection we could retrieve the audit messages and order them based on date.