DBD-864 - Add Logger to AppendContentService
This adds a standard Ruby Logger instance and swaps out the puts
statements. For now, there is no config for the logger; it just uses
STDOUT, so the behavior will be essentially identical to the puts
version (except for the one log line at debug level since the default
level is set to INFO).