Remove unnecessary top-level includes
Seeing "include SomeClassorModule" outside of any scope is surprising
in ruby land, and has a side-effect on the global namespace. These
were primarily functioning as "require" statements. We opted to delete
them instead of replacing them with "requires" because rails autoloading
is already loading the files, and the autoloading convention is being
followed project wide.