Move #tombstone logic to GenericWork#entomb!
The Hyrax::GenericWorksController#tombstone method reached into
the GenericWork to manipulate its values.

This commit moves the logic of what it means to tombstone a GenericWork
into the GenericWork class itself. The controller is then able to
delegate to the model with one line.

This is not a functional change.